refactor: edit import path and rearrage test files
